# services/catalog-service/src/main/kotlin/com/mundiapolis/library/catalog/adapter/in/web/CatalogReadController.kt
package com.mundiapolis.library.catalog.adapter.`in`.web

import com.mundiapolis.library.catalog.dto.CatalogSearchFilters
import com.mundiapolis.library.catalog.dto.CatalogSearchResult
import com.mundiapolis.library.catalog.dto.CatalogReviewPage
import com.mundiapolis.library.catalog.dto.Edition
import com.mundiapolis.library.catalog.dto.Work
import com.mundiapolis.library.catalog.dto.LearningResource
import com.mundiapolis.library.catalog.dto.LearningResourcePage
import com.mundiapolis.library.catalog.service.CatalogService
import com.mundiapolis.library.catalog.service.LearningResourceService
import org.springframework.http.HttpStatus
import org.springframework.security.access.prepost.PreAuthorize
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.RestController
import org.springframework.web.server.ResponseStatusException
import java.util.UUID

@RestController
@RequestMapping("/api/v1/catalog")
class CatalogReadController(
    private val catalogService: CatalogService,
    private val learningResources: LearningResourceService,
) {
    @GetMapping("/works/{workId}")
    @PreAuthorize("hasAuthority('SCOPE_catalog.read')")
    fun work(@PathVariable workId: UUID): Work = catalogService.getWork(workId.toString())
        ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Work not found")

    @GetMapping("/editions/{editionId}")
    @PreAuthorize("hasAuthority('SCOPE_catalog.read')")
    fun edition(@PathVariable editionId: UUID): Edition = catalogService.getEdition(editionId.toString())
        ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Edition not found")

    @GetMapping("/search")
    @PreAuthorize("hasAuthority('SCOPE_catalog.search')")
    fun search(
        @RequestParam(required = false) query: String?,
        @RequestParam(required = false) genre: String?,
        @RequestParam(required = false) authorId: String?,
        @RequestParam(required = false) availableOnly: Boolean?,
        @RequestParam(required = false) minRating: Double?,
        @RequestParam(required = false) sortBy: String?,
        @RequestParam(required = false) page: Int?,
        @RequestParam(required = false) limit: Int?,
    ): CatalogSearchResult = catalogService.searchCatalog(
        CatalogSearchFilters(
            query = query,
            genre = genre,
            authorId = authorId,
            availableOnly = availableOnly,
            minRating = minRating,
            sortBy = sortBy,
            page = page,
            limit = limit,
        ),
    )

    @GetMapping("/genres")
    @PreAuthorize("hasAuthority('SCOPE_catalog.search')")
    fun genres(): List<String> = catalogService.getDistinctGenres()

    @GetMapping("/works/{workId}/reviews")
    @PreAuthorize("hasAuthority('SCOPE_catalog.read')")
    fun reviews(
        @PathVariable workId: UUID,
        @RequestParam(required = false) page: Int?,
        @RequestParam(required = false) limit: Int?,
    ): CatalogReviewPage = catalogService.getPublishedReviews(workId.toString(), page, limit)

    @GetMapping("/learning-resources/{resourceId}")
    @PreAuthorize("hasAuthority('SCOPE_catalog.learning-resource.read')")
    fun learningResource(@PathVariable resourceId: UUID): LearningResource =
        learningResources.find(resourceId)
            ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "Learning resource not found")

    @GetMapping("/learning-resources")
    @PreAuthorize("hasAuthority('SCOPE_catalog.learning-resource.read')")
    fun learningResources(
        @RequestParam(required = false) query: String?,
        @RequestParam(required = false) category: String?,
        @RequestParam(required = false) page: Int?,
        @RequestParam(required = false) limit: Int?,
    ): LearningResourcePage = learningResources.search(query, category, page, limit)

    @GetMapping("/learning-resource-categories")
    @PreAuthorize("hasAuthority('SCOPE_catalog.learning-resource.read')")
    fun learningResourceCategories(): List<String> = learningResources.categories()
}