# services/catalog-service/src/main/resources/static/openapi/catalog-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mundiapolis Catalog API",
    "version": "1.5.0",
    "description": "Authoritative catalog metadata and privacy-safe review commands and reads. Copy counts and returned-loan eligibility are versioned projections owned by Circulation."
  },
  "paths": {
...
    "/api/v1/catalog/works/{workId}": {
      "get": {
        "operationId": "getWork",
        "x-required-scopes": ["catalog.read"],
        "parameters": [{ "$ref": "#/components/parameters/WorkId" }],
        "responses": {
          "200": { "description": "Work found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Work" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateWork",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateWorkRequest" } } } },
        "responses": {
          "200": { "description": "Work updated or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/editions/{editionId}": {
      "get": {
        "operationId": "getEdition",
        "x-required-scopes": ["catalog.read"],
        "parameters": [{ "$ref": "#/components/parameters/EditionId" }],
        "responses": {
          "200": { "description": "Edition found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Edition" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "operationId": "updateEdition",
        "x-required-scopes": ["catalog.manage"],
        "parameters": [
          { "$ref": "#/components/parameters/EditionId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEditionRequest" } } } },
        "responses": {
          "200": { "description": "Edition updated or exact command replayed", "headers": { "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/api/v1/catalog/works/{workId}/reviews": {
      "get": {
        "operationId": "getPublishedReviews",
        "x-required-scopes": ["catalog.read"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 0, "maximum": 10000, "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Published review page without member identifiers", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogReviewPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createReview",
        "x-required-scopes": ["catalog.review.write"],
        "parameters": [
          { "$ref": "#/components/parameters/WorkId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateReviewRequest" } } } },
        "responses": {
          "201": { "description": "Review created or exact command replayed", "headers": { "Location": { "description": "Canonical review command resource.", "schema": { "type": "string", "format": "uri-reference" } }, "ETag": { "description": "Quoted aggregate version.", "schema": { "type": "string" } }, "Idempotency-Replayed": { "schema": { "type": "boolean" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogCommandResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" }
        }
      }
    },
...
    "/api/v1/catalog/search": {
      "get": {
        "operationId": "searchCatalog",
        "x-required-scopes": ["catalog.search"],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          { "name": "genre", "in": "query", "schema": { "type": "string", "maxLength": 120 } },
          { "name": "authorId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "availableOnly", "in": "query", "schema": { "type": "boolean", "default": false } },
          { "name": "minRating", "in": "query", "schema": { "type": "number", "minimum": 0, "maximum": 5, "default": 0 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["title", "rating", "publicationYear"], "default": "title" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 0, "maximum": 10000, "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Filtered catalog page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CatalogSearchResult" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/catalog/genres": {
      "get": {
        "operationId": "getCatalogGenres",
        "x-required-scopes": ["catalog.search"],
        "responses": {
          "200": { "description": "Sorted active catalog genres", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/catalog/learning-resources": {
      "get": {
        "operationId": "searchLearningResources",
        "x-required-scopes": ["catalog.learning-resource.read"],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string", "maxLength": 200 } },
          { "name": "category", "in": "query", "schema": { "type": "string", "maxLength": 128 } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 0, "maximum": 10000, "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 24 } }
        ],
        "responses": {
          "200": { "description": "Active learning-resource page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LearningResourcePage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/catalog/learning-resources/{resourceId}": {
      "get": {
        "operationId": "getLearningResource",
        "x-required-scopes": ["catalog.learning-resource.read"],
        "parameters": [{ "name": "resourceId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Learning resource found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LearningResource" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/catalog/learning-resource-categories": {
      "get": {
        "operationId": "getLearningResourceCategories",
        "x-required-scopes": ["catalog.learning-resource.read"],
        "responses": {
          "200": { "description": "Sorted active learning-resource categories", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "WorkId": { "name": "workId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
      "EditionId": { "name": "editionId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 128, "pattern": "^[ -~]+$" } },
      "IfMatch": { "name": "If-Match", "in": "header", "required": true, "description": "Exact quoted aggregate version, for example \"0\".", "schema": { "type": "string", "pattern": "^\"(0|[1-9][0-9]*)\"$" } }
    },
    "responses": {
      "BadRequest": { "description": "Invalid request" },
      "Unauthorized": { "description": "Missing or invalid bearer token" },
      "Forbidden": { "description": "Required scope is missing" },
      "NotFound": { "description": "Resource not found" },
      "Conflict": { "description": "Aggregate, contributor metadata, or idempotency conflict" },
      "PayloadTooLarge": { "description": "Command body exceeds 128 KiB" },
      "UnprocessableEntity": { "description": "Returned-loan evidence does not authorize this review" }
    },
    "schemas": {
      "Author": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "bio": { "type": ["string", "null"] }
        }
      },
      "Work": {
        "type": "object",
        "additionalProperties": false,
        "required": ["workId", "title", "summary", "description", "genre", "rating", "authors"],
        "properties": {
          "workId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "description": { "type": "string" },
          "genre": { "type": "string" },
          "rating": { "type": "number", "minimum": 0, "maximum": 5 },
          "authors": { "type": "array", "items": { "$ref": "#/components/schemas/Author" } }
        }
      },
      "Edition": {
        "type": "object",
        "additionalProperties": false,
        "required": ["editionId", "workId", "title", "isbn", "publisher", "publicationYear", "language", "pageCount", "totalCopies", "availableCopies", "isActive"],
        "properties": {
          "editionId": { "type": "string", "format": "uuid" },
          "workId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "isbn": { "type": "string" },
          "publisher": { "type": "string" },
          "publicationYear": { "type": "integer" },
          "language": { "type": "string" },
          "pageCount": { "type": "integer", "minimum": 1 },
          "coverUrl": { "type": ["string", "null"], "format": "uri" },
          "coverColor": { "type": ["string", "null"], "pattern": "^#[0-9A-Fa-f]{6}$" },
          "videoUrl": { "type": ["string", "null"], "format": "uri" },
          "totalCopies": { "type": "integer", "minimum": 0, "description": "Disposable projection from Circulation." },
          "availableCopies": { "type": "integer", "minimum": 0, "description": "Disposable projection from Circulation." },
          "isActive": { "type": "boolean" }
        }
      },
      "CatalogSearchResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["editions", "total", "page", "totalPages"],
        "properties": {
          "editions": { "type": "array", "items": { "$ref": "#/components/schemas/Edition" } },
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
      "CatalogReview": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reviewId", "workId", "rating", "content", "reviewerLabel", "createdAt", "updatedAt"],
        "properties": {
          "reviewId": { "type": "string", "format": "uuid" },
          "workId": { "type": "string", "format": "uuid" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "content": { "type": "string", "minLength": 1, "maxLength": 4000 },
          "reviewerLabel": { "type": "string", "const": "Verified reader" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CatalogReviewPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reviews", "total", "page", "totalPages"],
        "properties": {
          "reviews": { "type": "array", "items": { "$ref": "#/components/schemas/CatalogReview" } },
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
      "CreateAuthorRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["contributorId", "name"],
        "properties": {
          "contributorId": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "maxLength": 300 },
          "bio": { "type": ["string", "null"], "maxLength": 5000 }
        }
      },
...
      "UpdateWorkRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "summary", "description", "genre", "authors", "reason"],
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "summary": { "type": "string", "maxLength": 1000 },
          "description": { "type": "string", "maxLength": 10000 },
          "genre": { "type": "string", "maxLength": 120 },
          "authors": { "type": "array", "minItems": 1, "maxItems": 20, "items": { "$ref": "#/components/schemas/CreateAuthorRequest" } },
          "reason": { "type": "string", "minLength": 8, "maxLength": 500 }
        }
      },
      "UpdateEditionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "isbn", "publisher", "publicationYear", "language", "pageCount", "reason"],
        "properties": {
          "title": { "type": "string", "maxLength": 500 },
          "isbn": { "type": "string", "maxLength": 32 },
          "publisher": { "type": "string", "maxLength": 300 },
          "publicationYear": { "type": "integer", "minimum": 1000, "maximum": 3000 },
          "language": { "type": "string", "maxLength": 80 },
          "pageCount": { "type": "integer", "minimum": 1, "maximum": 100000 },
          "coverUrl": { "type": ["string", "null"], "format": "uri", "maxLength": 2048 },
          "coverColor": { "type": ["string", "null"], "pattern": "^#[0-9A-Fa-f]{6}$" },
          "videoUrl": { "type": ["string", "null"], "format": "uri", "maxLength": 2048 },
          "reason": { "type": "string", "minLength": 8, "maxLength": 500 }
        }
      },
...
      "CreateReviewRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["rating", "content"],
        "properties": {
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "content": { "type": "string", "minLength": 1, "maxLength": 4000 }
        }
      },
...
      "LearningResource": {
        "type": "object",
        "additionalProperties": false,
        "required": ["resourceId", "title", "category", "language", "sourceName", "sourceUrl"],
        "properties": {
          "resourceId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "author": { "type": ["string", "null"] },
          "description": { "type": ["string", "null"] },
          "category": { "type": "string" },
          "language": { "type": "string" },
          "coverUrl": { "type": ["string", "null"], "format": "uri" },
          "coverAlt": { "type": ["string", "null"] },
          "sourceName": { "type": "string" },
          "sourceUrl": { "type": "string", "format": "uri" }
        }
      },
      "LearningResourcePage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["resources", "total", "page", "totalPages"],
        "properties": {
          "resources": { "type": "array", "items": { "$ref": "#/components/schemas/LearningResource" } },
          "total": { "type": "integer", "minimum": 0 },
          "page": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 }
        }
      },
...
      "CatalogCommandResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["aggregateType", "aggregateId", "aggregateVersion", "occurredAt"],
        "properties": {
          "aggregateType": { "type": "string", "enum": ["work", "edition", "review"] },
          "aggregateId": { "type": "string", "format": "uuid" },
          "aggregateVersion": { "type": "integer", "minimum": 0 },
          "occurredAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}